refactor(dashboard): avg, ops indicator decimal point correction

+++ src/main/java/com/dev/dugout/domain/user/dto/PlayerInsightDto.java
@@ -4,6 +4,8 @@ package com.dev.dugout.domain.user.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class PlayerInsightDto {
@@ -15,14 +17,14 @@ public class PlayerInsightDto {
 
 
     // 2026년 예상 핵심 지표
-    private Double predictedAvg;   // 예상 타율 (예: 0.312)
+    private BigDecimal predictedAvg;   // 예상 타율 (예: 0.312)
     private Integer predictedHr;   // 예상 홈런 (예: 25)
-    private Double predictedOps;   // 예상 OPS (예: 0.895)
+    private BigDecimal predictedOps;   // 예상 OPS (예: 0.895)
 
     // 2025년 대비 변화폭
-    private Double avgDiff;        // 타율 차이
+    private BigDecimal avgDiff;        // 타율 차이
     private Integer hrDiff;        // 홈런 차이
-    private Double opsDiff;        // OPS 차이
+    private BigDecimal opsDiff;        // OPS 차이
 
     // Bedrock이 생성한 요약 문구 또는 상세 인사이트 JSON
     private String insightSummary;
